fix: change auth schema files
