+++ server/services/auth.ts
@@ -1,10 +1,9 @@
 import bcrypt from 'bcrypt';
 import { User } from '../models';
 import { CustomError } from '../utils';
-import { signupSchema } from '../validations/schema';
 import { generateToken } from '../utils/jwt/generateToken';
 import { UserData } from '../interfaces/auth';
-import { loginSchema } from '../validations';
+import { signupSchema, loginSchema } from '../validations';
 import { userLoginAttrs } from '../interfaces/auth';
 
 const signupService = async (userData: UserData) => {
